MEL-214 Get the Search on Every Page
This is a quick refactor that moves lifecycle methods into the search box, but is otherwise mostly just rearranging the location of some of the search compents within the app structure.
