+++ src/Components/Search/Results/Result/ResultIndex/index.js
