+++ src/Components/Search/Results/Result/ResultThumbnail/index.js
