+++ src/Components/Search/Results/Result/index.js
