+++ src/Components/Search/ResultsView/style.css
